+++ stock-system/src/main/java/saffchen/command/CreateXlsFileCommand.java
@@ -1,0 +1,29 @@
+package saffchen.command;
+import com.google.api.client.auth.oauth2.Credential;
+import com.google.api.client.extensions.java6.auth.oauth2.AuthorizationCodeInstalledApp;
+import com.google.api.client.extensions.jetty.auth.oauth2.LocalServerReceiver;
+import com.google.api.client.googleapis.auth.oauth2.GoogleAuthorizationCodeFlow;
+import com.google.api.client.googleapis.auth.oauth2.GoogleClientSecrets;
+import com.google.api.client.googleapis.javanet.GoogleNetHttpTransport;
+import com.google.api.client.json.jackson2.JacksonFactory;
+import com.google.api.client.util.store.FileDataStoreFactory;
+import com.google.api.services.sheets.v4.Sheets;
+import com.google.api.services.sheets.v4.SheetsScopes;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.security.GeneralSecurityException;
+import java.util.List;
+
+public class CreateXlsFileCommand implements Command {
+    private ReceiverDB receiverDB;
+
+    public CreateXlsFileCommand(ReceiverDB receiverDB) {
+        this.receiverDB = receiverDB;
+    }
+
+    @Override
+    public void doCommand() throws GeneralSecurityException, IOException {
+        receiverDB.createXLS();
+    }
+}
