+++ stock-system/src/main/resources/credentials.json
@@ -1,0 +1,1 @@
+{"installed":{"client_id":"323778323757-tea93tgr8putnsrn6gsaa05ctd40trpd.apps.googleusercontent.com","project_id":"elite-epoch-350914","auth_uri":"https://accounts.google.com/o/oauth2/auth","token_uri":"https://oauth2.googleapis.com/token","auth_provider_x509_cert_url":"https://www.googleapis.com/oauth2/v1/certs","client_secret":"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","redirect_uris":["http://localhost"]}}
